requirements.txt
transformers
torch
pillow
scikit-learn
pandas
tqdm
numpy
matplotlib
